+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/ContributionProvidersBoxTestGenerated.java
@@ -277,6 +277,12 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
     }
 
     @Test
+    @TestMetadata("implicitClassKeyMapKey.kt")
+    public void testImplicitClassKeyMapKey() {
+      runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/implicitClassKeyMapKey.kt");
+    }
+
+    @Test
     @TestMetadata("intoMap.kt")
     public void testIntoMap() {
       runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/intoMap.kt");