+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -146,6 +146,12 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
       }
 
       @Test
+      @TestMetadata("PresenterFunctionImplicitReturnType.kt")
+      public void testPresenterFunctionImplicitReturnType() {
+        runTest("compiler-tests/src/test/data/diagnostic/api/circuit/PresenterFunctionImplicitReturnType.kt");
+      }
+
+      @Test
       @TestMetadata("PresenterFunctionInvalidParams.kt")
       public void testPresenterFunctionInvalidParams() {
         runTest("compiler-tests/src/test/data/diagnostic/api/circuit/PresenterFunctionInvalidParams.kt");
@@ -155,12 +161,6 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
       @TestMetadata("PresenterFunctionInvalidReturn.kt")
       public void testPresenterFunctionInvalidReturn() {
         runTest("compiler-tests/src/test/data/diagnostic/api/circuit/PresenterFunctionInvalidReturn.kt");
-      }
-
-      @Test
-      @TestMetadata("PresenterFunctionImplicitReturnType.kt")
-      public void testPresenterFunctionImplicitReturnType() {
-        runTest("compiler-tests/src/test/data/diagnostic/api/circuit/PresenterFunctionImplicitReturnType.kt");
       }
 
       @Test